add TaskButtons component
